# specs/001-phase2-kickoff/contracts/todo-api.openapi.yaml
openapi: 3.0.3
info:
  title: Phase II Todo Web API
  version: 0.2.0
  description: >
    Contract for authenticated multi-user task lifecycle APIs in Phase II.
servers:
  - url: https://api.example.com
security:
  - bearerAuth: []
paths:
  /tasks:
    get:
      summary: List tasks for authenticated user
      operationId: listTasks
      responses:
        "200":
          description: User task list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaskRead"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      summary: Create task for authenticated user
      operationId: createTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskCreate"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskRead"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /tasks/{taskId}:
    get:
      summary: Get task details
      operationId: getTask
      parameters:
        - $ref: "#/components/parameters/taskId"
      responses:
        "200":
          description: Task details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskRead"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      summary: Update task fields
      operationId: updateTask
      parameters:
        - $ref: "#/components/parameters/taskId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskRead"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      summary: Delete task
      operationId: deleteTask
      parameters:
        - $ref: "#/components/parameters/taskId"
      responses:
        "204":
          description: Deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  /tasks/{taskId}/toggle-complete:
    post:
      summary: Toggle completion status
      operationId: toggleComplete
      parameters:
        - $ref: "#/components/parameters/taskId"
      responses:
        "200":
          description: Task with toggled status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskRead"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    taskId:
      name: taskId
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
  schemas:
    TaskCreate:
      type: object
      required: [title]
      additionalProperties: false
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 120
        description:
          type: string
          maxLength: 500
          nullable: true
    TaskUpdate:
      type: object
      additionalProperties: false
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 120
        description:
          type: string
          maxLength: 500
          nullable: true
        is_completed:
          type: boolean
    TaskRead:
      type: object
      required: [id, title, is_completed, created_at, updated_at]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        is_completed:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      required: [error, code]
      properties:
        error:
          type: string
        code:
          type: string
  responses:
    BadRequest:
      description: Validation failure
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Task not found for authenticated user
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"